+++ R/rl_write_control_parameters_mt.R
@@ -5,6 +5,7 @@
 #'
 #' @param x - dataset containing the trading results for one trading robot
 #' @param path_control_files - path where control parameters will be saved
+#' @param num_trades_to_consider - number of last trades to use for RL modeling simulations, default value 100
 #'
 #' @details Function is used by the R script Adapt_RL_MT_control.R
 #'
@@ -12,12 +13,14 @@
 #'
 #' @export
 #'
-#' @author (C) 2019 Vladimir Zhbanko
+#' @author (C) 2019, 2021 Vladimir Zhbanko
 #'
 #' @examples
 #'
 #' \donttest{
 #' # test lasts 15 sec:
+#' dir <- normalizePath(tempdir(),winslash = "/")
+#'
 #' library(dplyr)
 #' library(readr)
 #' library(ReinforcementLearning)
@@ -26,11 +29,15 @@
 #' data(trading_systemDF)
 #'
 #' # use optimal control parameters found by auxiliary function
-#' rl_write_control_parameters_mt(trading_systemDF, path_control_files = tempfile())
+#' rl_write_control_parameters_mt(x = trading_systemDF,
+#'                                path_control_files = dir,
+#'                                num_trades_to_consider = 100)
 #' }
 #'
 #'
-rl_write_control_parameters_mt <- function(x, path_control_files){
+rl_write_control_parameters_mt <- function(x,
+                                           path_control_files,
+                                           num_trades_to_consider = 100){
 
   requireNamespace("dplyr", quietly = TRUE)
   requireNamespace("readr", quietly = TRUE)
@@ -38,7 +45,10 @@ rl_write_control_parameters_mt <- function(x, path_control_files){
 # delete DF_RES if it is exist
 if(exists("DF_RES")){rm(DF_RES)}
 
-# perform set of operations for every trading system
+  # only consider last 100 observations in the input dataset to make code running faster
+  x <- x %>%
+    dplyr::arrange(OrderCloseTime) %>%
+    tail(num_trades_to_consider)
 
 # sum(x$Profit)
 #states <- c("tradewin", "tradeloss")
@@ -50,9 +60,9 @@ actions <- c("ON", "OFF") # 'ON' and 'OFF' are referring to decision to trade wi
 #control <- list(alpha = 0.3, gamma = 0.6, epsilon = 0.1)
 
 # Create RL models for all the space
-Alpha <- seq(from = 0.1, to = 0.9, by = 0.2)
-Gamma <- seq(from = 0.1, to = 0.9, by = 0.2)
-Epsyl <- seq(from = 0.1, to = 0.9, by = 0.2)
+Alpha <- seq(from = 0.1, to = 0.9, by = 0.4)
+Gamma <- seq(from = 0.1, to = 0.9, by = 0.4)
+Epsyl <- seq(from = 0.1, to = 0.9, by = 0.4)
 
 for (ALF in Alpha) {
   for (GAM in Gamma) {
@@ -105,6 +115,7 @@ DF_RES1 <- DF_RES %>%
 # recording control parameters for this system (this function is intended to run 1x week)
 
   # create directory where to write files if not exists yet
+  path_control_files <- file.path(path_control_files, 'control')
   if(!dir.exists(path_control_files)){
     dir.create(path_control_files)
   }