+++ NEWS.md
@@ -14,6 +14,7 @@
 
 * added fail safe stops in function `check_if_optimize.R`
 * experimenting with time series statistical transformation adding functions `mt_stat_transf` and `mt_stat_evaluate`. Learning opportunity around clustering for automatic market type generation
+* added fail safe and change the way how to find control parameters for Reinforcement Learning to avoid slow code execution
 
 # lazytrade 0.4.3